requirements.txt
# === Core Frameworks ===
fastapi
uvicorn

# === Deep Learning ===
torch
torchvision
transformers
ultralytics

# === Image Processing ===
pillow
opencv-python

# === Server & Uploads ===
python-multipart
python-dotenv

# === AWS / S3 Support ===
boto3

# === Tracking / Metrics ===
mlflow
scikit-learn

prometheus-fastapi-instrumentator
